corrected packaging structure of endpoints and Refactored Types and metadata endpoints.
