--- src/main/java/uk/gov/legislation/endpoints/document/api/params/Monarch.java
@@ -1,13 +1,0 @@
-package uk.gov.legislation.endpoints.document.api.params;
-
-import io.swagger.v3.oas.annotations.Parameter;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.PARAMETER)
-@Parameter(description = "The name of a monarch, relative to which the year is given", example = "Eliz2")
-public @interface Monarch { }
