+++ src/main/java/uk/gov/legislation/params/Years.java
@@ -1,0 +1,13 @@
+package uk.gov.legislation.params;
+
+import io.swagger.v3.oas.annotations.Parameter;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.PARAMETER)
+@Parameter(description = "A year or range of years, relative to the monarch", example = "10-11")
+public @interface Years { }
